section1.4.2 completed.
