--- src/chapter01/chat/main.go
@@ -1,41 +1,0 @@
-package main
-
-import (
-	"flag"
-	"log"
-	"net/http"
-	"text/template"
-	"path/filepath"
-	"sync"
-)
-
-// templは１つのテンプレートを表します
-type templateHandler struct {
-	once		sync.Once
-	filename	string
-	templ		*template.Template
-}
-
-// ServerHTTPはHTTPリクエストを処理します
-func (t *templateHandler) ServeHTTP(w http.ResponseWriter, r *http.Request){
-	t.once.Do(func() {
-			t.templ = template.Must(template.ParseFiles(filepath.Join("templates", t.filename)))
-		})
-	t.templ.Execute(w, r)
-}
-
-func main() {
-	var addr = flag.String("addr", ":8080", "アプリケーションのアドレス")
-	flag.Parse()
-	r := newRoom()
-	http.Handle("/", &templateHandler{filename: "chat.html"})
-	http.Handle("/room", r)
-
-	go r.run()
-
-	// Webサーバを起動します
-	log.Println("Webサーバーを開始します。ポート：", *addr)
-	if err := http.ListenAndServe(*addr, nil); err != nil {
-		log.Fatal("ListenAndServe:", err)
-	}
-}
