--- src/chapter01/chat/templates/chat.html
@@ -1,49 +1,0 @@
-<html>
-	<head>
-		<title>チャット</title>
-		<style>
-			input { display: block; }
-			ul { list-style: none; }
-		</style>
-	</head>
-
-	<body>
-		<ul id="messages"></ul>
-		WebSocketを使ったチャットアプリケーション
-		<form id="chatbox">
-			<textarea></textarea>
-			<input type="submit" value="送信" />
-		</form>
-
-		<script src="//ajax.googleapis.com/ajax/libs/jquery/1.11.1/jquery.min.js"></script>
-		<script>
-			$(function(){
-				var socket = null;
-				var msgBox = $("#chatbox textarea");
-				var messages = $("#messages");
-				$("#chatbox").submit(function(){
-					if(!msgBox.val()) return false;
-					if(!socket){
-						alert("エラー：WebSocket接続が行われていません。");
-						return false;
-					}
-					socket.send(msgBox.val());
-					msgBox.val("");
-					return false;
-				});
-				if(!window["WebSocket"]){
-					alert("エラー：WebSocketに対応していないブラウザです。");
-				}else{
-					//socket = new WebSocket("ws://localhost:8080/room");
-					socket = new WebSocket("ws://{{.Host}}/room");
-					socket.onclose = function(){
-						alert("接続が終了しました。");
-					}
-					socket.onmessage = function(e){
-						messages.append($("<li>").text(e.data));
-					}
-				}
-			});
-		</script>
-	</body>
-</html>
