+++ src/chat/main.go
@@ -1,0 +1,44 @@
+package main
+
+import (
+	"flag"
+	"log"
+	"net/http"
+	"text/template"
+	"path/filepath"
+	"sync"
+	"os"
+	"trace"
+)
+
+// templは１つのテンプレートを表します
+type templateHandler struct {
+	once		sync.Once
+	filename	string
+	templ		*template.Template
+}
+
+// ServerHTTPはHTTPリクエストを処理します
+func (t *templateHandler) ServeHTTP(w http.ResponseWriter, r *http.Request){
+	t.once.Do(func() {
+			t.templ = template.Must(template.ParseFiles(filepath.Join("templates", t.filename)))
+		})
+	t.templ.Execute(w, r)
+}
+
+func main() {
+	var addr = flag.String("addr", ":8080", "アプリケーションのアドレス")
+	flag.Parse()
+	r := newRoom()
+	r.tracer = trace.New(os.Stdout)
+	http.Handle("/", &templateHandler{filename: "chat.html"})
+	http.Handle("/room", r)
+
+	go r.run()
+
+	// Webサーバを起動します
+	log.Println("Webサーバーを開始します。ポート：", *addr)
+	if err := http.ListenAndServe(*addr, nil); err != nil {
+		log.Fatal("ListenAndServe:", err)
+	}
+}
