+++ src/trace/tracer.go
@@ -1,0 +1,23 @@
+package trace
+
+import(
+	"fmt"
+	"io"
+)
+
+type Tracer interface {
+	Trace(...interface{})
+}
+
+func New(w io.Writer) Tracer {
+	return &tracer{out: w}
+}
+
+type tracer struct {
+	out io.Writer
+}
+
+func (t *tracer) Trace(a ...interface{}) {
+	t.out.Write([]byte(fmt.Sprint(a...)))
+	t.out.Write([]byte("\n"))
+}
